IMP optim: ellipsoidal certifier family with soundness gates
